fixed geodata and optimalizations
