+++ src/core/map/geodata-processor/worker-main.js
@@ -6,6 +6,7 @@ import {getLayer as getLayer_, getLayerPropertyValue as getLayerPropertyValue_,
 import {processLineStringPass as processLineStringPass_, processLineStringGeometry as processLineStringGeometry_} from './worker-linestring.js';
 import {processPointArrayPass as processPointArrayPass_, processPointArrayGeometry as processPointArrayGeometry_} from './worker-pointarray.js';
 import {processPolygonPass as processPolygonPass_} from './worker-polygon.js';
+import {postGroupMessage as postGroupMessage_, optimizeGroupMessages as optimizeGroupMessages_} from './worker-message.js';
 
 //get rid of compiler mess
 var globals = globals_;
@@ -18,6 +19,7 @@ var processPointArrayPass = processPointArrayPass_;
 var processPolygonPass = processPolygonPass_;
 var processLineStringGeometry = processLineStringGeometry_;
 var processPointArrayGeometry = processPointArrayGeometry_;
+var postGroupMessage = postGroupMessage_, optimizeGroupMessages = optimizeGroupMessages_;
 
 var exportedGeometries = [];
 var featureCache = new Array(1024), featureCacheIndex = 0, finalFeatureCache = new Array(1024), finalFeatureCacheIndex = 0, finalFeatureCacheIndex2 = 0;
@@ -306,7 +308,7 @@ function processGroup(group, lod) {
         bboxDelta[1] / bboxResolution,
         bboxDelta[2] / bboxResolution];
 
-    postMessage({'command':'beginGroup', 'id': group['id'], 'bbox': [bboxMin, bboxMax], 'origin': bboxMin});
+    postGroupMessage({'command':'beginGroup', 'id': group['id'], 'bbox': [bboxMin, bboxMax], 'origin': bboxMin});
 
     //process points
     var points = group['points'] || [];
@@ -323,11 +325,11 @@ function processGroup(group, lod) {
     globals.featureType = 'polygon';
     processFeatures('polygon', polygons, lod, 'polygon', groupId);
 
+    postGroupMessage({'command':'endGroup'});
+
     if (globals.groupOptimize) {
         optimizeGroupMessages();
     }
-
-    postMessage({'command':'endGroup'});
 }
 
 
@@ -358,140 +360,6 @@ function processGeodata(data, lod) {
     //console.log("processGeodata-ready");
 }
 
-function optimizeGroupMessages() {
-    //debugger;
-    //loop messages
-    var messages = globals.messageBuffer;
-    var j, lk, k, message2, job2, vbufferSize, vbuffer, index, buff, buff2;
-
-    for (var i = 0, li = globals.messageBufferIndex; i < li; i++) {
-        var message = messages[i];
-        var job = message.job;
-        var type = job['type'];
-        var signature = message.signature;
-        
-        if (!job['hitable'] && !message.reduced &&  //!job["culling"] &&
-            !(type == 'icon' || type == 'label' ||
-              type == 'line-geometry' || type == 'point-geometry')) {
-            
-            switch(type) {
-            case 'flat-line':
-                vbufferSize = job['vertexBuffer'].length;
-
-                for (j = i + 1; j < li; j++) {
-                    message2 = messages[j];
-                        
-                    if (message2.signature == signature) {
-                        message2.reduced = true;
-                        vbufferSize += message2.job['vertexBuffer'].length;
-                    }
-                }
-
-                vbuffer = new Float32Array(vbufferSize);
-                index = 0;
-
-                for (j = i; j < li; j++) {
-                    message2 = messages[j];
-                    job2 = message2.job;
-                        
-                    if (message2.signature == signature) {
-                        buff = job2['vertexBuffer'];
-                        job2['vertexBuffer'] = null;
-                        for (k = 0, lk = buff.length; k < lk; k++) {
-                            vbuffer[index+k] = buff[k];
-                        }
-                        index += lk;
-                    }
-                }
-
-                job['vertexBuffer'] = vbuffer;
-                message.arrays = [vbuffer.buffer];
-                break;
-                    
-            case 'pixel-line':
-            case 'line-label':
-            case 'flat-rline':
-
-                vbufferSize = job['vertexBuffer'].length;
-
-                for (j = i + 1; j < li; j++) {
-                    message2 = messages[j];
-
-                    if (message2.signature == signature) {
-                        message2.reduced = true;
-                        vbufferSize += message2.job['vertexBuffer'].length;
-                    }
-                }
-
-                vbuffer = new Float32Array(vbufferSize);
-                var nbuffer = new Float32Array(vbufferSize);
-                index = 0;
-
-                for (j = i; j < li; j++) {
-                    message2 = messages[j];
-                    job2 = message2.job;
-                        
-                    if (message2.signature == signature) {
-                        buff = job2['vertexBuffer'];
-                        job2['vertexBuffer'] = null;
-                            
-                        if (type == 'line-label') {
-                            buff2 = job2['texcoordsBuffer'];
-                            job2['texcoordsBuffer'] = null;
-                        } else {
-                            buff2 = job2['normalBuffer'];
-                            job2['normalBuffer'] = null;
-                        }
-                            
-                        for (k = 0, lk = buff.length; k < lk; k++) {
-                            vbuffer[index+k] = buff[k];
-                            nbuffer[index+k] = buff2[k];
-                        }
-
-                        index += lk;
-
-                        if (type == 'line-label') {
-                            var files = job['files'];
-                            var files2 = job2['files'];
-
-                            for (k = 0, lk = files2.length; k < lk; k++) {
-                                if (!files[k]) {
-                                    files[k] = [];
-                                }
-
-                                for (var m = 0, lm = files2[k].length; m < lm; m++) {
-                                    if (files[k].indexOf(files2[k][m]) == -1) {
-                                        files[k].push(files2[k][m]);
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-
-                job['vertexBuffer'] = vbuffer;
-
-                if (type == 'line-label') {
-                    job['texcoordsBuffer'] = nbuffer;
-                } else {
-                    job['normalBuffer'] = nbuffer;
-                }
-
-                message.arrays = [vbuffer.buffer, nbuffer.buffer];
-                break;
-            }
-
-            postMessage(message.job, message.arrays);
-            
-        } else if (!message.reduced) {
-
-            postMessage(message.job, message.arrays);
-
-        }
-    }
-
-    globals.messageBufferIndex = 0;
-} 
 
 self.onmessage = function (e) {
     var message = e.data;
@@ -517,6 +385,7 @@ self.onmessage = function (e) {
 
     case 'setFontMap':
         setFontMap(data);
+        postMessage({'command' : 'styleDone'});
         postMessage({'command' : 'ready'});
         break;
